requirements.txt
PyGithub
pandas
matplotlib
seaborn
fpdf2
python-dotenv
openai
google-genai
plotly
kaleido
streamlit
watchdog
